CMake Errors, conversion errors and types

+++ Space-Hunt/space_hunt_game.cpp
@@ -13,7 +13,7 @@ PlanetLevelScene planetLevel;
 WinScene win;
 
 short unsigned switchState;
-int volume;
+float volume;
 Vector2i resolution;
 
 int main() {
